feat: Bottombar
feat: Changes

feat: Colors
