+++ app/src/main/java/mx/pi5/localito/adapters/CategoryAdapter.java
@@ -1,4 +1,4 @@
-package mx.pi5.localito;
+package mx.pi5.localito.adapters;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -8,6 +8,8 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import mx.pi5.localito.R;
 
 public class CategoryAdapter extends RecyclerView.Adapter<CategoryAdapter.CategoryViewHolder> {
     private String[] categories;
