+++ app/src/main/java/mx/pi5/localito/adapters/ProductAdapter.java
@@ -1,4 +1,4 @@
-package mx.pi5.localito;
+package mx.pi5.localito.adapters;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,6 +7,8 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import mx.pi5.localito.R;
 
 public class ProductAdapter extends RecyclerView.Adapter<ProductAdapter.ProductViewHolder> {
     private String[] products;
