+++ app/src/main/java/mx/pi5/localito/fragments/HomeFragment.java
@@ -1,0 +1,37 @@
+package mx.pi5.localito.fragments;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import mx.pi5.localito.R;
+import mx.pi5.localito.databinding.FragmentHomeBinding;
+
+public class HomeFragment extends Fragment {
+    protected FragmentHomeBinding b;
+
+    public HomeFragment() {
+        // Required empty public constructor
+    }
+
+    public static HomeFragment newInstance() {
+        HomeFragment fragment = new HomeFragment();
+        return fragment;
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        b = FragmentHomeBinding.inflate(inflater, container, false);
+        return b.getRoot();
+    }
+}
