+++ app/src/main/res/layout/activity_stands_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".StandsList">
+    tools:context=".ProductCatalogue$StandsList">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvStands"
