+++ app/src/main/res/menu/bottombar_items.xml
@@ -1,4 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/home"
+        android:title="@string/home"
+        android:icon="@drawable/baseline_home_24"
+        app:showAsAction="always"
+        />
+    <item
+        android:id="@+id/categories"
+        android:title="@string/categories"
+        android:icon="@drawable/baseline_category_24"
+        app:showAsAction="always"
+        />
+    <item
+        android:id="@+id/user"
+        android:title="@string/user"
+        android:icon="@drawable/baseline_person_24"
+        app:showAsAction="always"
+        />
 </menu>
