+++ app/src/main/res/values/colors.xml
@@ -1,65 +1,64 @@
-<?xml version="1.0" encoding="utf-8"?>
 
 <resources>
-    <color name="seed">#FE7F2D</color>
-    <color name="md_theme_light_primary">#9C4400</color>
+    <color name="seed">#ff6939</color>
+    <color name="md_theme_light_primary">#AE3100</color>
     <color name="md_theme_light_onPrimary">#FFFFFF</color>
-    <color name="md_theme_light_primaryContainer">#FFDBCA</color>
-    <color name="md_theme_light_onPrimaryContainer">#331100</color>
-    <color name="md_theme_light_secondary">#406918</color>
+    <color name="md_theme_light_primaryContainer">#FFDBD0</color>
+    <color name="md_theme_light_onPrimaryContainer">#3A0A00</color>
+    <color name="md_theme_light_secondary">#77574D</color>
     <color name="md_theme_light_onSecondary">#FFFFFF</color>
-    <color name="md_theme_light_secondaryContainer">#C0F190</color>
-    <color name="md_theme_light_onSecondaryContainer">#0D2000</color>
-    <color name="md_theme_light_tertiary">#775A00</color>
+    <color name="md_theme_light_secondaryContainer">#FFDBD0</color>
+    <color name="md_theme_light_onSecondaryContainer">#2C150F</color>
+    <color name="md_theme_light_tertiary">#6B5E2F</color>
     <color name="md_theme_light_onTertiary">#FFFFFF</color>
-    <color name="md_theme_light_tertiaryContainer">#FFDF97</color>
-    <color name="md_theme_light_onTertiaryContainer">#251A00</color>
+    <color name="md_theme_light_tertiaryContainer">#F5E2A7</color>
+    <color name="md_theme_light_onTertiaryContainer">#231B00</color>
     <color name="md_theme_light_error">#BA1A1A</color>
     <color name="md_theme_light_errorContainer">#FFDAD6</color>
     <color name="md_theme_light_onError">#FFFFFF</color>
     <color name="md_theme_light_onErrorContainer">#410002</color>
-    <color name="md_theme_light_background">#FAFCFF</color>
-    <color name="md_theme_light_onBackground">#001F2A</color>
-    <color name="md_theme_light_surface">#FAFCFF</color>
-    <color name="md_theme_light_onSurface">#001F2A</color>
-    <color name="md_theme_light_surfaceVariant">#F4DED4</color>
-    <color name="md_theme_light_onSurfaceVariant">#52443D</color>
-    <color name="md_theme_light_outline">#85746B</color>
-    <color name="md_theme_light_inverseOnSurface">#E1F4FF</color>
-    <color name="md_theme_light_inverseSurface">#003547</color>
-    <color name="md_theme_light_inversePrimary">#FFB68F</color>
+    <color name="md_theme_light_background">#FFFBFF</color>
+    <color name="md_theme_light_onBackground">#201A18</color>
+    <color name="md_theme_light_surface">#FFFBFF</color>
+    <color name="md_theme_light_onSurface">#201A18</color>
+    <color name="md_theme_light_surfaceVariant">#F5DED7</color>
+    <color name="md_theme_light_onSurfaceVariant">#53433F</color>
+    <color name="md_theme_light_outline">#85736E</color>
+    <color name="md_theme_light_inverseOnSurface">#FBEEEA</color>
+    <color name="md_theme_light_inverseSurface">#362F2D</color>
+    <color name="md_theme_light_inversePrimary">#FFB59F</color>
     <color name="md_theme_light_shadow">#000000</color>
-    <color name="md_theme_light_surfaceTint">#9C4400</color>
-    <color name="md_theme_light_outlineVariant">#D7C2B9</color>
+    <color name="md_theme_light_surfaceTint">#AE3100</color>
+    <color name="md_theme_light_outlineVariant">#D8C2BC</color>
     <color name="md_theme_light_scrim">#000000</color>
-    <color name="md_theme_dark_primary">#FFB68F</color>
-    <color name="md_theme_dark_onPrimary">#542100</color>
-    <color name="md_theme_dark_primaryContainer">#773200</color>
-    <color name="md_theme_dark_onPrimaryContainer">#FFDBCA</color>
-    <color name="md_theme_dark_secondary">#A4D577</color>
-    <color name="md_theme_dark_onSecondary">#1B3700</color>
-    <color name="md_theme_dark_secondaryContainer">#2A5000</color>
-    <color name="md_theme_dark_onSecondaryContainer">#C0F190</color>
-    <color name="md_theme_dark_tertiary">#F1C03C</color>
-    <color name="md_theme_dark_onTertiary">#3E2E00</color>
-    <color name="md_theme_dark_tertiaryContainer">#5A4300</color>
-    <color name="md_theme_dark_onTertiaryContainer">#FFDF97</color>
+    <color name="md_theme_dark_primary">#FFB59F</color>
+    <color name="md_theme_dark_onPrimary">#5E1600</color>
+    <color name="md_theme_dark_primaryContainer">#852300</color>
+    <color name="md_theme_dark_onPrimaryContainer">#FFDBD0</color>
+    <color name="md_theme_dark_secondary">#E7BDB1</color>
+    <color name="md_theme_dark_onSecondary">#442A22</color>
+    <color name="md_theme_dark_secondaryContainer">#5D4037</color>
+    <color name="md_theme_dark_onSecondaryContainer">#FFDBD0</color>
+    <color name="md_theme_dark_tertiary">#D8C68D</color>
+    <color name="md_theme_dark_onTertiary">#3A2F05</color>
+    <color name="md_theme_dark_tertiaryContainer">#524619</color>
+    <color name="md_theme_dark_onTertiaryContainer">#F5E2A7</color>
     <color name="md_theme_dark_error">#FFB4AB</color>
     <color name="md_theme_dark_errorContainer">#93000A</color>
     <color name="md_theme_dark_onError">#690005</color>
     <color name="md_theme_dark_onErrorContainer">#FFDAD6</color>
-    <color name="md_theme_dark_background">#001F2A</color>
-    <color name="md_theme_dark_onBackground">#BFE9FF</color>
-    <color name="md_theme_dark_surface">#001F2A</color>
-    <color name="md_theme_dark_onSurface">#BFE9FF</color>
-    <color name="md_theme_dark_surfaceVariant">#52443D</color>
-    <color name="md_theme_dark_onSurfaceVariant">#D7C2B9</color>
-    <color name="md_theme_dark_outline">#9F8D84</color>
-    <color name="md_theme_dark_inverseOnSurface">#001F2A</color>
-    <color name="md_theme_dark_inverseSurface">#BFE9FF</color>
-    <color name="md_theme_dark_inversePrimary">#9C4400</color>
+    <color name="md_theme_dark_background">#201A18</color>
+    <color name="md_theme_dark_onBackground">#EDE0DC</color>
+    <color name="md_theme_dark_surface">#201A18</color>
+    <color name="md_theme_dark_onSurface">#EDE0DC</color>
+    <color name="md_theme_dark_surfaceVariant">#53433F</color>
+    <color name="md_theme_dark_onSurfaceVariant">#D8C2BC</color>
+    <color name="md_theme_dark_outline">#A08C87</color>
+    <color name="md_theme_dark_inverseOnSurface">#201A18</color>
+    <color name="md_theme_dark_inverseSurface">#EDE0DC</color>
+    <color name="md_theme_dark_inversePrimary">#AE3100</color>
     <color name="md_theme_dark_shadow">#000000</color>
-    <color name="md_theme_dark_surfaceTint">#FFB68F</color>
-    <color name="md_theme_dark_outlineVariant">#52443D</color>
+    <color name="md_theme_dark_surfaceTint">#FFB59F</color>
+    <color name="md_theme_dark_outlineVariant">#53433F</color>
     <color name="md_theme_dark_scrim">#000000</color>
 </resources>
