+++ app/src/main/res/values/themes.xml
@@ -1,31 +1,31 @@
+
 <resources xmlns:tools="http://schemas.android.com/tools">
-    <!-- Base application theme. -->
-        <style name="Theme.Localito" parent="Theme.Material3.Light.NoActionBar">
-            <item name="colorPrimary">@color/md_theme_light_primary</item>
-            <item name="colorOnPrimary">@color/md_theme_light_onPrimary</item>
-            <item name="colorPrimaryContainer">@color/md_theme_light_primaryContainer</item>
-            <item name="colorOnPrimaryContainer">@color/md_theme_light_onPrimaryContainer</item>
-            <item name="colorSecondary">@color/md_theme_light_secondary</item>
-            <item name="colorOnSecondary">@color/md_theme_light_onSecondary</item>
-            <item name="colorSecondaryContainer">@color/md_theme_light_secondaryContainer</item>
-            <item name="colorOnSecondaryContainer">@color/md_theme_light_onSecondaryContainer</item>
-            <item name="colorTertiary">@color/md_theme_light_tertiary</item>
-            <item name="colorOnTertiary">@color/md_theme_light_onTertiary</item>
-            <item name="colorTertiaryContainer">@color/md_theme_light_tertiaryContainer</item>
-            <item name="colorOnTertiaryContainer">@color/md_theme_light_onTertiaryContainer</item>
-            <item name="colorError">@color/md_theme_light_error</item>
-            <item name="colorErrorContainer">@color/md_theme_light_errorContainer</item>
-            <item name="colorOnError">@color/md_theme_light_onError</item>
-            <item name="colorOnErrorContainer">@color/md_theme_light_onErrorContainer</item>
-            <item name="android:colorBackground">@color/md_theme_light_background</item>
-            <item name="colorOnBackground">@color/md_theme_light_onBackground</item>
-            <item name="colorSurface">@color/md_theme_light_surface</item>
-            <item name="colorOnSurface">@color/md_theme_light_onSurface</item>
-            <item name="colorSurfaceVariant">@color/md_theme_light_surfaceVariant</item>
-            <item name="colorOnSurfaceVariant">@color/md_theme_light_onSurfaceVariant</item>
-            <item name="colorOutline">@color/md_theme_light_outline</item>
-            <item name="colorOnSurfaceInverse">@color/md_theme_light_inverseOnSurface</item>
-            <item name="colorSurfaceInverse">@color/md_theme_light_inverseSurface</item>
-            <item name="colorPrimaryInverse">@color/md_theme_light_inversePrimary</item>
-        </style>
+    <style name="Theme.Localito" parent="Theme.Material3.Light.NoActionBar">
+        <item name="colorPrimary">@color/md_theme_light_primary</item>
+        <item name="colorOnPrimary">@color/md_theme_light_onPrimary</item>
+        <item name="colorPrimaryContainer">@color/md_theme_light_primaryContainer</item>
+        <item name="colorOnPrimaryContainer">@color/md_theme_light_onPrimaryContainer</item>
+        <item name="colorSecondary">@color/md_theme_light_secondary</item>
+        <item name="colorOnSecondary">@color/md_theme_light_onSecondary</item>
+        <item name="colorSecondaryContainer">@color/md_theme_light_secondaryContainer</item>
+        <item name="colorOnSecondaryContainer">@color/md_theme_light_onSecondaryContainer</item>
+        <item name="colorTertiary">@color/md_theme_light_tertiary</item>
+        <item name="colorOnTertiary">@color/md_theme_light_onTertiary</item>
+        <item name="colorTertiaryContainer">@color/md_theme_light_tertiaryContainer</item>
+        <item name="colorOnTertiaryContainer">@color/md_theme_light_onTertiaryContainer</item>
+        <item name="colorError">@color/md_theme_light_error</item>
+        <item name="colorErrorContainer">@color/md_theme_light_errorContainer</item>
+        <item name="colorOnError">@color/md_theme_light_onError</item>
+        <item name="colorOnErrorContainer">@color/md_theme_light_onErrorContainer</item>
+        <item name="android:colorBackground">@color/md_theme_light_background</item>
+        <item name="colorOnBackground">@color/md_theme_light_onBackground</item>
+        <item name="colorSurface">@color/md_theme_light_surface</item>
+        <item name="colorOnSurface">@color/md_theme_light_onSurface</item>
+        <item name="colorSurfaceVariant">@color/md_theme_light_surfaceVariant</item>
+        <item name="colorOnSurfaceVariant">@color/md_theme_light_onSurfaceVariant</item>
+        <item name="colorOutline">@color/md_theme_light_outline</item>
+        <item name="colorOnSurfaceInverse">@color/md_theme_light_inverseOnSurface</item>
+        <item name="colorSurfaceInverse">@color/md_theme_light_inverseSurface</item>
+        <item name="colorPrimaryInverse">@color/md_theme_light_inversePrimary</item>
+    </style>
 </resources>
